Netlify creates functions with any file in root.  Moving app_settings/logger to utils folder
